Rename router into routes
