+++ packages/desmos-std/src/subspaces/mod.rs
@@ -1,3 +1,3 @@
 pub mod models;
 pub mod query;
-pub mod router;
+pub mod routes;
